+++ webpack.config.js
@@ -1,19 +1,19 @@
-const path = require("path");
+const path = require('path');
 
 module.exports = {
-  entry: "./src/extension.js",
-  target: "node",
-  mode: "production",
-  output: {
-    path: path.resolve(__dirname, "dist"),
-    filename: "extension.js",
-    libraryTarget: "commonjs2",
-  },
-  externals: {
-    vscode: "commonjs vscode",
-  },
-  resolve: {
-    extensions: [".js"],
-  },
+    entry: './src/extension.js',
+    target: 'node',
+    mode: 'production',
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'extension.js',
+        libraryTarget: 'commonjs2',
+    },
+    externals: {
+        vscode: 'commonjs vscode',
+    },
+    resolve: {
+        extensions: ['.js'],
+    },
 };